+++ fuc/api/pyvcf.py
@@ -909,18 +909,6 @@ class VcfFrame:
         return list(self.df.CHROM.unique())
 
     @property
-    def variants(self):
-        """list : List of variants."""
-        def one_row(r):
-            result = []
-            for alt in r.ALT.split(','):
-                result.append(f'{r.CHROM}-{r.POS}-{r.REF}-{alt}')
-            return ','.join(result)
-        s = self.df.apply(one_row, axis=1)
-        s = ','.join(s)
-        return s.split(',')
-
-    @property
     def sites_only(self):
         """bool : Whether the VCF is sites-only."""
         return not self.samples or 'FORMAT' not in self.df.columns
@@ -5449,3 +5437,41 @@ class VcfFrame:
             af = float(field.split(',')[j+1])
 
         return af
+
+    def variants(self):
+        """
+        List variants in VcfFrame.
+
+        Returns
+        -------
+        list
+            List of variants.
+
+        Examples
+        --------
+
+        >>> from fuc import pyvcf
+        >>> data = {
+        ...     'CHROM': ['chr1', 'chr2'],
+        ...     'POS': [100, 101],
+        ...     'ID': ['.', '.'],
+        ...     'REF': ['G', 'T'],
+        ...     'ALT': ['A', 'A,C'],
+        ...     'QUAL': ['.', '.'],
+        ...     'FILTER': ['.', '.'],
+        ...     'INFO': ['.', '.'],
+        ...     'FORMAT': ['GT', 'GT'],
+        ...     'A': ['0/1', '1/2']
+        ... }
+        >>> vf = pyvcf.VcfFrame.from_dict([], data)
+        >>> vf.variants()
+        ['chr1-100-G-A', 'chr2-101-T-A', 'chr2-101-T-C']
+        """
+        def one_row(r):
+            result = []
+            for alt in r.ALT.split(','):
+                result.append(f'{r.CHROM}-{r.POS}-{r.REF}-{alt}')
+            return ','.join(result)
+        s = self.df.apply(one_row, axis=1)
+        s = ','.join(s)
+        return s.split(',')